Use uint8_t for dimensions.

+++ include/fontlib.h
@@ -22,8 +22,8 @@ struct glyph_t
 
 struct font_t
 {
-    uint16_t start_row() const { return -min_y; }
-    uint16_t line_spacing() const { return 1 + max_y - min_y; }
+    uint8_t start_row() const { return -min_y; }
+    uint8_t line_spacing() const { return 1 + max_y - min_y; }
 
     uint8_t         height;
     int8_t          min_y;
